added mechanics and molecular
